+++ src/test/java/com/aquila/chess/strategy/mcts/NNTest.java
@@ -9,7 +9,7 @@ import java.util.*;
 public class NNTest implements INN {
 
     double step = 0.000001;
-    private double value = 0.1;
+    private float value = 0.1F;
 
     private final Map<Integer, Double> offsets = new HashMap<>();
 
@@ -49,7 +49,7 @@ public class NNTest implements INN {
     }
 
     @Override
-    public void fit(double[][][][] inputs, double[][] policies, double[][] values) {
+    public void fit(float[][][][] inputs, float[][] policies, float[][] values) {
 
     }
 
@@ -69,16 +69,16 @@ public class NNTest implements INN {
     }
 
     @Override
-    public synchronized List<OutputNN> outputs(double[][][][] nbIn, int len) {
+    public synchronized List<OutputNN> outputs(float[][][][] nbIn, int len) {
         List<OutputNN> ret = new ArrayList<>();
         for (int i = 0; i < len; i++) {
             this.value += step;
-            double[] policies = new double[PolicyUtils.MAX_POLICY_INDEX];
-            Arrays.fill(policies, 0.2);
+            float[] policies = new float[PolicyUtils.MAX_POLICY_INDEX];
+            Arrays.fill(policies, 0.2F);
             for (Map.Entry<Integer, Double> entry : this.offsets.entrySet()) {
                 policies[entry.getKey()] += entry.getValue();
             }
-            ret.add(new OutputNN(nbIn[i][InputsNNFactory.PLANE_COLOR][0][0] == 1.0 ? value + 0.2 : value, policies));
+            ret.add(new OutputNN(nbIn[i][InputsNNFactory.PLANE_COLOR][0][0] == 1.0F ? value + 0.2F : value, policies));
         }
         return ret;
     }
